backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
openai>=1.55.0
httpx>=0.27.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
python-docx==1.1.0
waitress==2.1.2
